requirements.txt
mysql-connector-python
pillow
flask
